security: eliminate 6 reports of CWE-78 in tests
NOW   CWEs = 4  (1 src,  3 tests); Severity: 0.4.0.0  Confidence: 0.1.0.3

PREV  CWEs = 10 (1 src,  9 tests); Severity: 0.10.0.0 Confidence: 0.1.0.9
