+++ tests/biskotaki_ci/snapshot/biskotaki_ci_no_input/test_lint_passes.py
@@ -13,9 +13,8 @@ import pytest
         'biskotaki-gold-standard',
     ],
 )
-def test_running_lint_passes(snapshot_name, test_root):
+def test_running_lint_passes(snapshot_name, my_run_subprocess, test_root):
     """Verify Snapshot Project passes `tox -e lint` out of the box."""
-    import subprocess
     from pathlib import Path
 
     # LINT Biskotaki
@@ -25,10 +24,11 @@ def test_running_lint_passes(snapshot_name, test_root):
     assert snapshot_dir.is_dir()
 
     # Programmatically run Lint, with the entrypoint we suggest, for a Dev to run
-    res = subprocess.run(  # tox -e lint
-        [sys.executable, '-m', 'tox', '-r', '-vv', '-e', 'lint'],
+    res = my_run_subprocess(  # tox -e lint
+        *[sys.executable, '-m', 'tox', '-vv', '-e', 'lint'],
         cwd=snapshot_dir,
         check=False,  # prevent raising exception, so we can do clean up
+        shell=False,  # prevent execution of untrusted input
     )
 
     # SANITY verify .tox/lint exists, when using tox 3.x
@@ -40,7 +40,10 @@ def test_running_lint_passes(snapshot_name, test_root):
     shutil.rmtree(snapshot_dir / '.tox' / 'lint')
 
     # Check that Code passes Lint out of the box
-    assert res.returncode == 0
+    assert (
+        res.exit_code == 0
+    ), f"Failed to run `tox -e lint` for {snapshot_name}\nSTDOUT:\n{res.stdout}\nSTDERR:\n{res.stderr}"
+
 
 
 @pytest.mark.slow
@@ -53,9 +56,8 @@ def test_running_lint_passes(snapshot_name, test_root):
         'biskotaki-gold-standard',
     ],
 )
-def test_running_ruff_passes(snapshot_name, test_root):
+def test_running_ruff_passes(snapshot_name, my_run_subprocess, test_root):
     """Verify Snapshot Project passes `tox -e ruff` out of the box."""
-    import subprocess
     from pathlib import Path
 
     # Load Snapshot
@@ -64,11 +66,16 @@ def test_running_ruff_passes(snapshot_name, test_root):
     assert snapshot_dir.is_dir()
 
     # Programmatically run Lint, with the entrypoint we suggest, for a Dev to run
-    res = subprocess.run(  # tox -e ruff
-        [sys.executable, '-m', 'tox', '-r', '-vv', '-e', 'ruff'],
+    res = my_run_subprocess(  # tox -e ruff
+        *[sys.executable, '-m', 'tox', '-vv', '-e', 'ruff'],
         cwd=snapshot_dir,
         check=False,  # prevent raising exception, so we can do clean up
+        shell=False,  # prevent execution of untrusted input
     )
+    # SANITY verify .tox/ruff exists, when using tox 3.x
+    assert (
+        res.exit_code == 0
+    ), f"Failed to run `tox -e ruff` for {snapshot_name}\nSTDOUT:\n{res.stdout}\nSTDERR:\n{res.stderr}"
 
     if (snapshot_dir / '.tox' / 'ruff').exists():
         # Remove .tox/ruff folder, created by tox 3.x
@@ -76,9 +83,5 @@ def test_running_ruff_passes(snapshot_name, test_root):
 
         shutil.rmtree(snapshot_dir / '.tox' / 'ruff')
 
-    # SANITY verify .tox/ruff exists, when using tox 3.x
-    assert (
-        res.returncode == 0
-    ), f"Failed to run `tox -e ruff` for {snapshot_name}\nSTDOUT:\n{res.stdout.decode(encoding='utf-8')}\nSTDERR:\n{res.stderr.decode(encoding='utf-8')}"
 
     # VERIFIED that Generator emits python code that pass Ruff out of the box !!
